refactor: split oversized modules into focused submodules
- explore/_session.py 1212→952 (+ _env, _pathutils, _policies)
- explore/repl.py 895→146 (+ _commands, _render, _completion)
- monitor/viewer.py 837→237 (+ _sources, _widgets)
- explore/_codegen.py 676→501 (+ _plans)
- docs/html.py 538→295 (+ _styles)

Pure mechanical moves; every original module keeps compat
re-exports so no importer changed. Suite: 620 passed, 2 skipped.
